Add MACD Indicator

+++ Albedo/Models/IndicatorData.cs
@@ -5,9 +5,9 @@ namespace Albedo.Models
     public class IndicatorData
     {
         public DateTime Date { get; set; }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
 
-        public IndicatorData(DateTime date, decimal value)
+        public IndicatorData(DateTime date, double value)
         {
             Date = date;
             Value = value;
